Sprint_7_some_fix

+++ helpers.py
@@ -6,8 +6,9 @@ import allure
 
 class Delivery:
 
+    @staticmethod
     @allure.step("Генератор инфо для регистрации доставщика")
-    def generation_register_data_delivery_man(self):
+    def generation_register_data_delivery_man():
         def generation_random_string(length):
             letters = string.ascii_lowercase
             random_string = ''.join(random.choice(letters) for i in range(length))
@@ -21,15 +22,16 @@ class Delivery:
 
         return payload
 
+    @staticmethod
     @allure.step("Регистрация доставщика")
-    def registration_delivery_man(self):
-        payload = Delivery.generation_register_data_delivery_man
-        response = requests.post(Url.create_delivery_man, data=payload)
-        return payload
+    def registration_delivery_man(payload):  # Передаем payload в качестве аргумента
+        response = requests.post(Url.create_delivery_man, json=payload)  # Использование json вместо data
+        return response
 
+    @staticmethod
     @allure.step("Удаление инфо о доставщике")
-    def remove_delivery_man(data):
-        payload = data
-        respons = requests.post(Url.authorization_delivery_man, data=payload)
-        delivery_man_id = respons.json()["id"]
-        requests.delete(f'{Url.create_delivery_man}/{delivery_man_id}')
+    def remove_delivery_man(payload):  # Добавлен параметр payload
+        response = requests.post(Url.authorization_delivery_man, json=payload)  # Использование json вместо data
+        delivery_man_id = response.json().get("id")
+        if delivery_man_id:  # Проверяем наличие ID перед отправкой запроса на удаление
+            requests.delete(f'{Url.create_delivery_man}/{delivery_man_id}')
